+++ src/ast/passes/semantic_analyser.cpp
@@ -1102,6 +1102,20 @@ void SemanticAnalyser::visit(Call &call)
     }
     call.type = CreateUInt64();
     call.type.SetAS(AddrSpace::kernel);
+  } else if (call.func == "percpu_kaddr") {
+    if (check_varargs(call, 1, 2)) {
+      check_arg(call, Type::string, 0, true);
+      if (call.vargs.size() == 2)
+        check_arg(call, Type::integer, 1, false);
+
+      auto symbol = bpftrace_.get_string_literal(call.vargs.at(0));
+      if (bpftrace_.btf_->get_var_type(symbol).IsNoneTy()) {
+        LOG(ERROR, call.loc, err_)
+            << "Could not resolve variable \"" << symbol << "\" from BTF";
+      }
+    }
+    call.type = CreateUInt64();
+    call.type.SetAS(AddrSpace::kernel);
   } else if (call.func == "uaddr") {
     auto probe = get_probe(call.loc, call.func);
     if (probe == nullptr)
@@ -1672,6 +1686,10 @@ void SemanticAnalyser::validate_map_key(const SizedType &key,
     LOG(ERROR, loc, err_) << "context cannot be used as a map key";
   }
 
+  if (key.IsHistTy() || key.IsLhistTy() || key.IsStatsTy()) {
+    LOG(ERROR, loc, err_) << key << " cannot be used as a map key";
+  }
+
   if (is_final_pass() && key.IsNoneTy()) {
     LOG(ERROR, loc, err_) << "Invalid expression for assignment: ";
   }
@@ -1982,6 +2000,12 @@ void SemanticAnalyser::binop_ptr(Binop &binop)
       binop.type = CreateInt(64);
     else
       invalid_op();
+  }
+  // Might need an additional pass to resolve the type
+  else if (other.IsNoneTy()) {
+    if (is_final_pass()) {
+      invalid_op();
+    }
   }
   // Binop on a pointer and something else
   else {